Archive fruit stuff from template
