+++ client/components/archived/SelectedFruit.tsx
@@ -1,8 +1,8 @@
-import { Fruit } from '../../models/fruit.ts'
+import { Fruit } from '../../../models/fruit.ts'
 
 import { useState } from 'react'
 
-import { GridForm, ColOne, ColTwoText, Button } from './Styled.tsx'
+import { GridForm, ColOne, ColTwoText, Button } from '../Styled.tsx'
 
 interface Props {
   fruit: Fruit
